// src/app/movies/model/movie.model.ts
//import { Rate } from './rate.model'

export interface Movie {
    id: number
    name: string
    sinopse: string
    category: string
    createdBy: string
    imagePath: string
    createdAt: string
}

// export class Mov implements Rate {

//     private id: number
//     private name: string
//     private sinopse: string
//     private createdBy: string
//     private imagePath: string
//     private createdAt: Date    
    
//     iconPath: string    
//     comments: string    
    

//     constructor(id: number, name: string, sinopose: string, createdBy: string, imagePath: string) {
//         this.id = id
//         this.name = name
//         this.sinopse = sinopose
//         this.createdBy = createdBy
//         this.imagePath = imagePath
//         this.createdAt = new Date
        
//     }
// }